Move web routes into routes.py as a FastAPI APIRouter
Completes the main.py split. All endpoints (dashboard, TestFlight ID and
Apprise URL management, control, config editor) plus the Pydantic request
models and the Jinja2 templates move into routes.py as an APIRouter that
imports from state/config/utils. main.py creates the FastAPI app, keeps the
app-level auth dependency, and does app.include_router(router).

app_start_time moves to state.py (used by the routes). Pruned the now
route-only imports out of main.py (pyflakes-clean); also fixed a latent
missing `import threading` surfaced in routes.py (restart Timer) and a
redundant top-level `import re` in state.py. API tests updated to patch the
add-ID validation on `routes`.

main.py: ~1,010 -> ~470 lines (from ~2,180 before the whole split).
